# openapi/chef-swagger-specs-v1.yaml
openapi: 3.0.1
info:
  title: YOLO Chef API
  description: YOLO Chef API Specification
  version: 0.0.1
servers:
  - url: /
paths:
  /api/v1/users:
    post:
      tags:
        - users
      summary: Create a user
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        required: true
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User created successfully
                   
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/{username}/userProfiles:
    post:
      tags:
        - users
      summary: Create user profile
      operationId: createUserProfile
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileRequest'
        required: true
      responses:
        '201':
          description: User profile created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User profile created successfully
                   
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/ideas/{idea_id}/recipes:
    post:
      tags:
        - recipes
      summary: Create a recipe based on a specific idea by the chef
      operationId: createRecipe
      parameters:
        - name: idea_id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        description: Recipe details to be created
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RecipeRequest'
        required: true
      responses:
        '201':
          description: Recipe created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Recipe created successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - recipes
      summary: Get all recipes by chef for an idea
      operationId: getRecipesForAnIdeaByChef
      parameters:
        - name: idea_id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  recipes:
                    type: array
                    items:
                      $ref: '#/components/schemas/RecipeSummary'
        '403':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/recipes/{recipe_id}:
    get:
      tags:
        - recipes
      summary: Get the recipe details by recipe ID for the chef
      operationId: getRecipeDetails
      parameters:
        - name: recipe_id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Recipe details fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeDetailSummary'
        '403':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
               
    patch:
      tags:
        - recipes
      summary: Update recipe status
      operationId: updateRecipeStatus
      parameters:
        - name: recipe_id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: The new status for the recipe.
              required:
                - status
        required: true
      responses:
        '201':
          description: Recipe updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Recipe status updated successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
    UserProfileRequest:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        contact_number:
          type: string
        house:
          type: string
        street:
          type: string
        area:
          type: string
        zip_code:
          type: string
        city:
          type: string
        country:
          type: string
        currency_code:
          type: string
...
    RecipeRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 20
        description:
          type: string
          maxLength: 255
        price:
          type: integer
          format: int64
        serving_size:
          type: integer
       
        images:
          type: array
          items:
            type: string
            format: binary
          description: "Array of image files to upload"
                     
      required:
        - title
        - description
        - price
        - serving_size
    RecipeSummary:
      type: object
      properties:
        id:
          type: integer
          format: int32
        recipe_name:
          type: string
        description:
          type: string
        serving_size:
          type: integer
        price:
          type: integer
          format: int64
        idea_title:
          type: string
        images:
          type: array
          items:
              type: object
              properties:
                  url:
                    type: string
                    example: "https://example.com/images/spaghetti-carbonara.jpg"
      required:
        - title
        - description
        - serving_size
        - price
        - idea_title
        - status
        - images
    RecipeDetailSummary:
      type: object
      properties:
        recipe:
          type: object
          properties:
            recipe_name:
              type: string
            description:
              type: string
            serving_size:
              type: integer
            price:
              type: integer
              format: int64
            idea_title:
              type: string
            status:
              type: string
            created_at:
              type: string
              format: DATETIME
            images:
              type: array
              items:
                type: object
                properties:
                  url:
                    type: string
                    example: "https://example.com/images/spaghetti-carbonara.jpg"
      required:
        - title
        - description
        - serving_size
        - price
        - idea_title
        - status
        - images
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: A brief message describing the error.
        details:
          type: string
          description: Additional details about the error.